feat: more splits
